+++ src/application.rs
@@ -3,6 +3,7 @@
 
 use clap::{Parser, ValueEnum};
 
+use crate::checkers::angularjs::AngularJSChecker;
 use crate::checkers::dovecot::DovecotChecker;
 use crate::checkers::exim::EximChecker;
 use crate::checkers::gsap::GsapChecker;
@@ -66,12 +67,13 @@ impl Application {
             Box::new(NginxChecker::new()),
             Box::new(PHPChecker::new()),
             Box::new(OpenSSLChecker::new()),
+            Box::new(AngularJSChecker::new()),
+            Box::new(GsapChecker::new()),
             Box::new(JQueryChecker::new()),
             Box::new(HandlebarsChecker::new()),
             Box::new(LodashChecker::new()),
             Box::new(PhpMyAdminChecker::new()),
             Box::new(Typo3Checker::new()),
-            Box::new(GsapChecker::new()),
         ];
 
         trace!("Returning the Application");