+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- AngularJS detection
 - GSAP detection
 - TYPO3 detection
 - phpMyAdmin detection